refactoring subfolders and file names for NumPy frontend.
