+++ ivy/functional/frontends/numpy/mathematical_functions/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa
+from . import arithmetic_operations
+from .arithmetic_operations import *
+from . import trigonometric_functions
+from .trigonometric_functions import *
