+++ ivy/functional/frontends/numpy/mathematical_functions/arithmetic_operations.py
@@ -24,25 +24,3 @@ def add(
 
 
 add.unsupported_dtypes = {"torch": ("float16",)}
-
-
-def tan(
-    x,
-    /,
-    out=None,
-    *,
-    where=True,
-    casting="same_kind",
-    order="k",
-    dtype=None,
-    subok=True,
-):
-    if dtype:
-        x = ivy.astype(ivy.array(x), ivy.as_ivy_dtype(dtype))
-    ret = ivy.tan(x, out=out)
-    if ivy.is_array(where):
-        ret = ivy.where(where, ret, ivy.default(out, ivy.zeros_like(ret)), out=out)
-    return ret
-
-
-tan.unsupported_dtypes = {"torch": ("float16",)}
