+++ ivy/functional/frontends/numpy/mathematical_functions/trigonometric_functions.py
@@ -1,0 +1,24 @@
+# global
+import ivy
+
+
+def tan(
+    x,
+    /,
+    out=None,
+    *,
+    where=True,
+    casting="same_kind",
+    order="k",
+    dtype=None,
+    subok=True,
+):
+    if dtype:
+        x = ivy.astype(ivy.array(x), ivy.as_ivy_dtype(dtype))
+    ret = ivy.tan(x, out=out)
+    if ivy.is_array(where):
+        ret = ivy.where(where, ret, ivy.default(out, ivy.zeros_like(ret)), out=out)
+    return ret
+
+
+tan.unsupported_dtypes = {"torch": ("float16",)}
